refactor(db): move trigger-maintenance tests + apply style consistency
WHY: Batch G cluster 2 of 3. T22/T40-T48 + trigger_sync_on_* +
multi-location-rename + soft-delete/restore + tag-rename-propagates land
in search_triggers.rs. Helpers consumed from common/mod.rs.

Also folds in 3 style NITs from Task 3 code review:
- //! before #![allow] in search_semantics.rs (matches dominant convention).
- Inline WHY on #![allow] attrs in common/mod.rs (matches Task 2 style).
- No blank line between use common and use perima_core (single import group).

Adds #[allow(dead_code, unused_imports)] to src/search_repo.rs::tests to
silence -D warnings on helpers now only used by the 2 remaining proptests;
Task 5 deletes the mod wholesale.

Test count unchanged (118).
